refactored common/Icons
