+++ src/components/ContentPages/Dashboard/DashboardItems/StatusCard.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { makeStyles } from "@material-ui/core/styles";
 import { Paper, Typography, Slide, SvgIcon} from "@material-ui/core";
-import { dataStatusIcons } from "../../../common/DataIcons/DataIcons";
+import { dataStatusIcons } from "../../../common/Icons";
 
 
 const useStyles = makeStyles(theme => ({
