+++ src/components/common/Icons/dataLangList.js
@@ -1,0 +1,14 @@
+
+  import ru from "./data/ru.png";
+  import ua from "./data/ua.png";
+  import us from "./data/us.png";
+
+  const dataLangList = { ru: ru, ua: ua, eng: us };
+
+  export default dataLangList;
+  
+  
+    
+  
+  
+
