+++ src/components/common/Icons/dataListIcons.js
@@ -1,0 +1,32 @@
+import {
+  Language,
+  Memory,
+  GetApp,
+  ImportExport,
+  FiberPin,
+  Input,
+  Router,
+  SettingsEthernet,
+  Dns
+} from "@material-ui/icons";
+
+const dataListIcons = {
+  cpu: Memory,
+  loader: GetApp,
+  minerva: ImportExport,
+  pin: FiberPin,
+  hdmi: Input,
+  ExternalIP: Language,
+  IpAddress: Router,
+  DefaultGateway: SettingsEthernet,
+  DnsServer1: Dns,
+  DnsServer2: Dns
+};
+
+export default dataListIcons; 
+  
+  
+    
+  
+  
+
