+++ src/components/common/Icons/dataRoutsIcons.js
@@ -1,0 +1,34 @@
+import {
+  Dashboard,
+  BarChart,
+  AccountTree,
+  ExitToApp,
+  EmojiEvents,
+  Accessible,
+  PeopleAlt,
+  Settings,
+  Home,
+  MeetingRoom,
+  AssignmentInd
+} from "@material-ui/icons";
+
+const dataRoutsIcons = {
+  home: Home,
+  dashboard: Dashboard,
+  reports: BarChart,
+  about: AccountTree,
+  old: Accessible,
+  users: PeopleAlt,
+  auth: ExitToApp,
+  contacts: EmojiEvents,
+  profile: AssignmentInd,
+  logout: MeetingRoom,
+  settings: Settings
+};
+  
+export default  dataRoutsIcons;
+  
+    
+  
+  
+
