+++ src/components/common/Icons/dataStatusIcons.js
@@ -1,0 +1,25 @@
+
+  import {
+    mdiTelevisionBox,
+    mdiPowerSettings,
+    mdiWan,
+    mdiSettingsBox,
+    mdiCellphone
+  } from "@mdi/js";
+  
+    const dataStatusIcons = {
+      model: mdiTelevisionBox,
+      online: mdiWan,
+      power: mdiPowerSettings,
+      version: mdiSettingsBox,
+      mobile: mdiCellphone
+    };
+  
+  
+    export default dataStatusIcons; 
+  
+  
+    
+  
+  
+
