+++ src/components/common/Icons/dataUsersRoleIcons.js
@@ -1,0 +1,17 @@
+import {
+  SupervisedUserCircle,
+  AccountCircle,
+  LocalCafe
+} from "@material-ui/icons";
+
+const dataUsersRoleIcons = {
+  3: LocalCafe,
+  2: SupervisedUserCircle,
+  1: AccountCircle
+};
+  
+export default dataUsersRoleIcons;  
+    
+  
+  
+
