+++ src/components/common/Icons/index.js
@@ -1,0 +1,5 @@
+export { default as dataLangList } from "./dataLangList.js";
+export { default as dataListIcons } from "./dataListIcons.js";
+export { default as dataRoutsIcons } from "./dataRoutsIcons.js";
+export { default as dataStatusIcons } from "./dataStatusIcons.js";
+export { default as dataUsersRoleIcons } from "./dataUsersRoleIcons.js";
