+++ src/components/common/ProgressLines/index.js
@@ -1,0 +1,2 @@
+export { default as MainLinearProgress } from './MainProgressLine.js';
+export { default as SearchLinearProgress } from './SearchProgressLine.js';
